backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
Werkzeug==3.0.1
requests==2.31.0
psutil==5.9.6
